refactor(vt): extract kitty graphics snapshots
Kitty placement, image, and unicode-placeholder rules were split across snapshot extraction and a placeholder helper. That made the VT Snapshot grid path responsible for renderer-adjacent graphics policy and kept placeholder text suppression far from the rest of Kitty handling.

Keep the external VT Snapshot and FrameSource seams unchanged while concentrating those rules in one VT Core-internal Module. Snapshot extraction can stay focused on grid, cursor, modes, and dirty state, and future Kitty fixes have one test surface.

No tracked issue.
