+++ crates/seance-vt/src/kitty_graphics.rs
@@ -1,143 +1,24 @@
-//! libghostty-vt snapshot extraction used by VT Core.
-
 use libghostty_vt::RenderState;
 use libghostty_vt::Terminal as VtTerminal;
 use libghostty_vt::kitty::graphics as kg;
-use libghostty_vt::render::{CellIteration, CellIterator, CursorVisualStyle, Dirty, RowIterator};
-use libghostty_vt::style::{self, PaletteIndex, RgbColor};
-use libghostty_vt::terminal::Mode;
+use libghostty_vt::render::{CellIteration, CellIterator, RowIterator};
+use libghostty_vt::style::{self, PaletteIndex};
 
-use crate::core::VtCoreError;
-use crate::frame::{
-    CellAttrs, CellColor, CursorInfo, CursorShape, DirtySnapshot, ImageInfo, ImageVisitor,
-    PlacementLayer, PlacementSnapshot, PlacementVisitor,
-};
-use crate::kitty_placeholder::{PLACEHOLDER_CP, diacritic_index};
-use crate::modes::TerminalModes;
-use crate::selection::GridPos;
-use crate::snapshot::{SnapshotImage, VtSnapshot};
+use crate::frame::{ImageInfo, ImageVisitor, PlacementLayer, PlacementSnapshot, PlacementVisitor};
+use crate::snapshot::SnapshotImage;
 
-/// Copy live libghostty mode flags into séance-owned input state.
-pub(crate) fn terminal_modes(vt: &VtTerminal<'static, 'static>) -> TerminalModes {
-    let mode = |m| vt.mode(m).unwrap_or(false);
-    TerminalModes {
-        cursor_keys: mode(Mode::DECCKM),
-        mouse_tracking: vt.is_mouse_tracking().unwrap_or(false),
-        mouse_format_sgr: mode(Mode::SGR_MOUSE),
-        bracketed_paste: mode(Mode::BRACKETED_PASTE),
-    }
+#[derive(Default)]
+pub(crate) struct KittyGraphicsSnapshot {
+    pub(crate) placements: Vec<PlacementSnapshot>,
+    pub(crate) images: Vec<SnapshotImage>,
 }
 
-pub(crate) struct SnapshotExtraction {
-    pub(crate) snapshot: VtSnapshot,
-    pub(crate) dirty_delta: DirtySnapshot,
-}
-
-pub(crate) fn extract_snapshot(
+pub(crate) fn extract_kitty_graphics(
     vt: &mut VtTerminal<'static, 'static>,
-    render_state: &mut RenderState<'static>,
     cell_width_px: u32,
     cell_height_px: u32,
-) -> Result<SnapshotExtraction, VtCoreError> {
-    let cols = vt.cols().unwrap_or(80);
-    let rows = vt.rows().unwrap_or(24);
-    let modes = terminal_modes(vt);
-
-    let mut out = VtSnapshot::empty(cols, rows);
-    out.modes = modes;
-
-    let dirty_delta;
-    {
-        let render_snapshot = render_state
-            .update(vt)
-            .map_err(|_| VtCoreError::libghostty("render state update"))?;
-        let global_dirty = render_snapshot.dirty().ok();
-        let mut dirty_rows = Vec::new();
-
-        let visible = render_snapshot.cursor_visible().unwrap_or(true);
-        let pos =
-            render_snapshot
-                .cursor_viewport()
-                .ok()
-                .flatten()
-                .map_or(GridPos::default(), |vp| GridPos {
-                    col: vp.x,
-                    row: vp.y,
-                });
-        let shape = render_snapshot
-            .cursor_visual_style()
-            .ok()
-            .and_then(map_cursor_shape);
-        out.cursor = CursorInfo {
-            pos,
-            visible,
-            wide: false,
-            shape,
-        };
-
-        let mut rows_iter =
-            RowIterator::new().map_err(|_| VtCoreError::libghostty("row iterator new"))?;
-        let mut cells_iter =
-            CellIterator::new().map_err(|_| VtCoreError::libghostty("cell iterator new"))?;
-        let mut row_iter = rows_iter
-            .update(&render_snapshot)
-            .map_err(|_| VtCoreError::libghostty("row iterator update"))?;
-        let mut scratch = String::with_capacity(4);
-
-        for row_idx in 0..rows {
-            let Some(row) = row_iter.next() else {
-                for _ in 0..cols {
-                    out.push_empty_cell();
-                }
-                continue;
-            };
-            match global_dirty {
-                Some(Dirty::Partial) if row.dirty().unwrap_or(true) => dirty_rows.push(row_idx),
-                None if row.dirty().unwrap_or(false) => dirty_rows.push(row_idx),
-                _ => {}
-            }
-            let mut cell_iter = cells_iter
-                .update(row)
-                .map_err(|_| VtCoreError::libghostty("cell iterator update"))?;
-            for _col in 0..cols {
-                let Some(cell) = cell_iter.next() else {
-                    out.push_empty_cell();
-                    continue;
-                };
-                scratch.clear();
-                if let Ok(graphs) = cell.graphemes() {
-                    scratch.extend(graphs);
-                }
-                if scratch.starts_with('\u{10EEEE}') {
-                    scratch.clear();
-                }
-                let style = cell.style().ok();
-                out.push_cell(
-                    &scratch,
-                    resolve_fg(cell, style.as_ref()),
-                    resolve_bg(cell, style.as_ref()),
-                    cell_attrs(style.as_ref()),
-                );
-            }
-            row.set_dirty(false)
-                .map_err(|_| VtCoreError::libghostty("clear row dirty"))?;
-        }
-
-        dirty_delta = match global_dirty {
-            Some(Dirty::Clean) => DirtySnapshot::Clean,
-            Some(Dirty::Full) => DirtySnapshot::Full,
-            Some(Dirty::Partial) | None => {
-                if dirty_rows.is_empty() {
-                    DirtySnapshot::Clean
-                } else {
-                    DirtySnapshot::Partial(dirty_rows)
-                }
-            }
-        };
-        render_snapshot
-            .set_dirty(Dirty::Clean)
-            .map_err(|_| VtCoreError::libghostty("clear render dirty"))?;
-    }
+) -> KittyGraphicsSnapshot {
+    let mut out = KittyGraphicsSnapshot::default();
 
     let mut placement_collector = SnapshotPlacementCollector::default();
     for layer in [
@@ -160,11 +41,57 @@ pub(crate) fn extract_snapshot(
     let _ = walk_images(vt, &mut image_collector);
     out.images = image_collector.images;
 
-    Ok(SnapshotExtraction {
-        snapshot: out,
-        dirty_delta,
-    })
+    out
 }
+
+pub(crate) fn is_placeholder_text(text: &str) -> bool {
+    text.starts_with('\u{10EEEE}')
+}
+
+/// The unicode placeholder base codepoint. Any cell whose first grapheme
+/// codepoint matches this is a Kitty virtual-placement placeholder.
+const PLACEHOLDER_CP: u32 = 0x10EEEE;
+
+/// Return the 0-based index of `cp` in the Kitty row/col diacritic
+/// alphabet, or `None` if `cp` is not a valid diacritic.
+fn diacritic_index(cp: u32) -> Option<u32> {
+    DIACRITICS.binary_search(&cp).ok().map(|i| i as u32)
+}
+
+/// Sorted diacritic codepoints used by the Kitty graphics protocol to
+/// encode row/column/image-id-high values.
+///
+/// Derived verbatim from:
+///   <https://sw.kovidgoyal.net/kitty/_downloads/f0a0de9ec8d9ff4456206db8e0814937/rowcolumn-diacritics.txt>
+///
+/// The index in this array is the encoded value.
+const DIACRITICS: &[u32] = &[
+    0x0305, 0x030D, 0x030E, 0x0310, 0x0312, 0x033D, 0x033E, 0x033F, 0x0346, 0x034A, 0x034B, 0x034C,
+    0x0350, 0x0351, 0x0352, 0x0357, 0x035B, 0x0363, 0x0364, 0x0365, 0x0366, 0x0367, 0x0368, 0x0369,
+    0x036A, 0x036B, 0x036C, 0x036D, 0x036E, 0x036F, 0x0483, 0x0484, 0x0485, 0x0486, 0x0487, 0x0592,
+    0x0593, 0x0594, 0x0595, 0x0597, 0x0598, 0x0599, 0x059C, 0x059D, 0x059E, 0x059F, 0x05A0, 0x05A1,
+    0x05A8, 0x05A9, 0x05AB, 0x05AC, 0x05AF, 0x05C4, 0x0610, 0x0611, 0x0612, 0x0613, 0x0614, 0x0615,
+    0x0616, 0x0617, 0x0657, 0x0658, 0x0659, 0x065A, 0x065B, 0x065D, 0x065E, 0x06D6, 0x06D7, 0x06D8,
+    0x06D9, 0x06DA, 0x06DB, 0x06DC, 0x06DF, 0x06E0, 0x06E1, 0x06E2, 0x06E4, 0x06E7, 0x06E8, 0x06EB,
+    0x06EC, 0x0730, 0x0732, 0x0733, 0x0735, 0x0736, 0x073A, 0x073D, 0x073F, 0x0740, 0x0741, 0x0743,
+    0x0745, 0x0747, 0x0749, 0x074A, 0x07EB, 0x07EC, 0x07ED, 0x07EE, 0x07EF, 0x07F0, 0x07F1, 0x07F3,
+    0x0816, 0x0817, 0x0818, 0x0819, 0x081B, 0x081C, 0x081D, 0x081E, 0x081F, 0x0820, 0x0821, 0x0822,
+    0x0823, 0x0825, 0x0826, 0x0827, 0x0829, 0x082A, 0x082B, 0x082C, 0x082D, 0x0951, 0x0953, 0x0954,
+    0x0F82, 0x0F83, 0x0F86, 0x0F87, 0x135D, 0x135E, 0x135F, 0x17DD, 0x193A, 0x1A17, 0x1A75, 0x1A76,
+    0x1A77, 0x1A78, 0x1A79, 0x1A7A, 0x1A7B, 0x1A7C, 0x1B6B, 0x1B6D, 0x1B6E, 0x1B6F, 0x1B70, 0x1B71,
+    0x1B72, 0x1B73, 0x1CD0, 0x1CD1, 0x1CD2, 0x1CDA, 0x1CDB, 0x1CE0, 0x1DC0, 0x1DC1, 0x1DC3, 0x1DC4,
+    0x1DC5, 0x1DC6, 0x1DC7, 0x1DC8, 0x1DC9, 0x1DCB, 0x1DCC, 0x1DD1, 0x1DD2, 0x1DD3, 0x1DD4, 0x1DD5,
+    0x1DD6, 0x1DD7, 0x1DD8, 0x1DD9, 0x1DDA, 0x1DDB, 0x1DDC, 0x1DDD, 0x1DDE, 0x1DDF, 0x1DE0, 0x1DE1,
+    0x1DE2, 0x1DE3, 0x1DE4, 0x1DE5, 0x1DE6, 0x1DFE, 0x20D0, 0x20D1, 0x20D4, 0x20D5, 0x20D6, 0x20D7,
+    0x20DB, 0x20DC, 0x20E1, 0x20E7, 0x20E9, 0x20F0, 0x2CEF, 0x2CF0, 0x2CF1, 0x2DE0, 0x2DE1, 0x2DE2,
+    0x2DE3, 0x2DE4, 0x2DE5, 0x2DE6, 0x2DE7, 0x2DE8, 0x2DE9, 0x2DEA, 0x2DEB, 0x2DEC, 0x2DED, 0x2DEE,
+    0x2DEF, 0x2DF0, 0x2DF1, 0x2DF2, 0x2DF3, 0x2DF4, 0x2DF5, 0x2DF6, 0x2DF7, 0x2DF8, 0x2DF9, 0x2DFA,
+    0x2DFB, 0x2DFC, 0x2DFD, 0x2DFE, 0x2DFF, 0xA66F, 0xA67C, 0xA67D, 0xA6F0, 0xA6F1, 0xA8E0, 0xA8E1,
+    0xA8E2, 0xA8E3, 0xA8E4, 0xA8E5, 0xA8E6, 0xA8E7, 0xA8E8, 0xA8E9, 0xA8EA, 0xA8EB, 0xA8EC, 0xA8ED,
+    0xA8EE, 0xA8EF, 0xA8F0, 0xA8F1, 0xAAB0, 0xAAB2, 0xAAB3, 0xAAB7, 0xAAB8, 0xAABE, 0xAABF, 0xAAC1,
+    0xFE20, 0xFE21, 0xFE22, 0xFE23, 0xFE24, 0xFE25, 0xFE26, 0x10A0F, 0x10A38, 0x1D185, 0x1D186,
+    0x1D187, 0x1D188, 0x1D189, 0x1D1AA, 0x1D1AB, 0x1D1AC, 0x1D1AD, 0x1D242, 0x1D243, 0x1D244,
+];
 
 #[derive(Default)]
 struct SnapshotPlacementCollector {
@@ -190,60 +117,6 @@ impl ImageVisitor for SnapshotImageCollector {
             height: info.height,
             rgba: info.rgba.to_vec(),
         });
-    }
-}
-
-fn resolve_bg(cell: &CellIteration<'_, '_>, style: Option<&style::Style>) -> CellColor {
-    if let Ok(Some(rgb)) = cell.bg_color() {
-        return rgb_to_cell_color(rgb);
-    }
-    style.map_or(CellColor::Default, |style| {
-        style_to_cell_color(&style.bg_color)
-    })
-}
-
-fn resolve_fg(cell: &CellIteration<'_, '_>, style: Option<&style::Style>) -> CellColor {
-    if let Ok(Some(rgb)) = cell.fg_color() {
-        return rgb_to_cell_color(rgb);
-    }
-    style.map_or(CellColor::Default, |style| {
-        style_to_cell_color(&style.fg_color)
-    })
-}
-
-fn cell_attrs(style: Option<&style::Style>) -> CellAttrs {
-    CellAttrs {
-        bold: style.is_some_and(|style| style.bold),
-        italic: style.is_some_and(|style| style.italic),
-        faint: style.is_some_and(|style| style.faint),
-        inverse: style.is_some_and(|style| style.inverse),
-        invisible: style.is_some_and(|style| style.invisible),
-    }
-}
-
-fn rgb_to_cell_color(c: RgbColor) -> CellColor {
-    CellColor::Rgb(c.r, c.g, c.b)
-}
-
-fn style_to_cell_color(sc: &style::StyleColor) -> CellColor {
-    match sc {
-        style::StyleColor::None => CellColor::Default,
-        style::StyleColor::Palette(PaletteIndex(idx)) => CellColor::Palette(*idx),
-        style::StyleColor::Rgb(rgb) => CellColor::Rgb(rgb.r, rgb.g, rgb.b),
-    }
-}
-
-// `BlockHollow` is ghostty's unfocused-window rendering, not a DECSCUSR
-// request — collapse into `Block`. Window-focus hollow cursors are a
-// separate feature tracked elsewhere. Unknown future variants (the
-// enum is `#[non_exhaustive]`) return `None` so the app falls back to
-// the user's configured shape.
-fn map_cursor_shape(s: CursorVisualStyle) -> Option<CursorShape> {
-    match s {
-        CursorVisualStyle::Bar => Some(CursorShape::Bar),
-        CursorVisualStyle::Block | CursorVisualStyle::BlockHollow => Some(CursorShape::Block),
-        CursorVisualStyle::Underline => Some(CursorShape::Underline),
-        _ => None,
     }
 }
 
@@ -721,5 +594,38 @@ fn cell_fg_to_id24(cell: &CellIteration<'_, '_>) -> u32 {
             let b = rgb.b as u32;
             (r << 16) | (g << 8) | b
         }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn alphabet_is_sorted() {
+        for pair in DIACRITICS.windows(2) {
+            assert!(pair[0] < pair[1], "not sorted at {:#x}", pair[0]);
+        }
+    }
+
+    #[test]
+    fn alphabet_has_297_entries() {
+        assert_eq!(DIACRITICS.len(), 297);
+    }
+
+    #[test]
+    fn placeholder_text_detects_base_codepoint() {
+        assert!(is_placeholder_text("\u{10EEEE}"));
+        assert!(is_placeholder_text("\u{10EEEE}\u{0305}"));
+        assert!(!is_placeholder_text(""));
+        assert!(!is_placeholder_text("x\u{10EEEE}"));
+    }
+
+    #[test]
+    fn spot_checks_match_upstream() {
+        assert_eq!(diacritic_index(0x0305), Some(0));
+        assert_eq!(diacritic_index(0x0483), Some(30));
+        assert_eq!(diacritic_index(0x1D244), Some(296));
+        assert_eq!(diacritic_index(0x0000), None);
     }
 }
